simplify hierarchy a bit
