+++ paper-server/src/main/java/org/bukkit/craftbukkit/event/CraftEventFactory.java
@@ -13,7 +13,6 @@ import io.papermc.paper.connection.PlayerConnection;
 import io.papermc.paper.event.block.BlockLockCheckEvent;
 import io.papermc.paper.event.connection.PlayerConnectionValidateLoginEvent;
 import io.papermc.paper.event.entity.BlockPlaceEntityEvent;
-import io.papermc.paper.event.entity.ItemSpawnEntityEvent;
 import io.papermc.paper.event.entity.ItemTransportingEntityValidateTargetEvent;
 import io.papermc.paper.event.player.PlayerBedFailEnterEvent;
 import io.papermc.paper.event.player.PlayerToggleEntityAgeLockEvent;
